requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
dj-database-url==1.2.0
Django==4.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-filter==22.1
django-jazzmin==2.6.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-spectacular==0.26.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
pyotp==2.8.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
